Add HtmlOutputTest along with test refactorings
