+++ tests/Constraint/ValidCreditorInformationPaymentReferenceCombinationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Sprain\SwissQrBill\Tests\Constraints;
+namespace Sprain\Tests\SwissQrBill\Constraints;
 
 use Sprain\SwissQrBill\Constraint\ValidCreditorInformationPaymentReferenceCombination;
 use Sprain\SwissQrBill\Constraint\ValidCreditorInformationPaymentReferenceCombinationValidator;
